MANTA-849 want more alarms for marlin agent

+++ npm/preuninstall.sh
@@ -6,8 +6,19 @@
 #
 
 set -o xtrace
-set -o errexit
+set -o pipefail
+. $(dirname $0)/../tools/common.sh
 
+mpi_probegroup=$(amon_config | json probegroup.name)
+[[ -n "$mpi_probegroup" ]] || fail "bad amon probegroup name"
+
+#
+# If this fails because amon is not available, drive on.  We're likely doing an
+# upgrade anyway.
+#
+$m_root/tools/amon_remove.sh "$mpi_probegroup"
+
+set -o errexit
 svcadm disable -s marlin-agent
 svccfg delete marlin-agent 
 rm -f $npm_config_smfdir/marlin-agent.xml
